# Improve performance and reduce the allocate when enforcing