--- a/Casbin.Benchmark/TestResource.cs
+++ b/Casbin.Benchmark/TestResource.cs
@@ -4,18 +4,12 @@
     {
         public TestResource(string name, string owner)
         {
-            this.name = name;
-            this.owner = owner;
+            Name = name;
+            Owner = owner;
         }
 
-#pragma warning disable IDE1006 // 命名样式
-        // ReSharper disable once InconsistentNaming
-        // ReSharper disable once MemberCanBePrivate.Global
-        public string name { get; set; }
+        public string Name { get; set; }
 
-        // ReSharper disable once InconsistentNaming
-        // ReSharper disable once MemberCanBePrivate.Global
-        public string owner { get; set; }
-#pragma warning restore IDE1006 // 命名样式
+        public string Owner { get; set; }
     }
 }