--- a/Casbin.Benchmark/examples/abac_model.conf
+++ b/Casbin.Benchmark/examples/abac_model.conf
@@ -8,4 +8,4 @@ p = sub, obj, act
 e = some(where (p.eft == allow))
 
 [matchers]
-m = r.sub == r.obj.owner
+m = r.sub == r.obj.Owner