--- a/NetCasbin/Abstractions/IExpressionHandler.cs
+++ b/NetCasbin/Abstractions/IExpressionHandler.cs
@@ -1,0 +1,26 @@
+﻿using System.Collections.Generic;
+using DynamicExpresso;
+
+namespace NetCasbin.Abstractions
+{
+    internal interface IExpressionHandler
+    {
+        public IDictionary<string, int> RequestTokens { get; }
+
+        public IDictionary<string, int> PolicyTokens { get; }
+
+        public IDictionary<string, Parameter> Parameters { get; } 
+
+        public void SetFunction(string name, AbstractFunction function);
+
+        public void SetGFunctions();
+
+        public void EnsureCreated(string expressionString, IReadOnlyList<object> requestValues);
+
+        public bool Invoke(string expressionString, IReadOnlyList<object> requestValues);
+
+        public void SetRequestParameters(IReadOnlyList<object> requestValues);
+
+        public void SetPolicyParameters(IReadOnlyList<string> policyValues);
+    }
+}