--- a/NetCasbin/Effect/IEffector.cs
+++ b/NetCasbin/Effect/IEffector.cs
@@ -1,6 +1,4 @@
-﻿using System;
-
-namespace NetCasbin.Effect
+﻿namespace NetCasbin.Effect
 {
     /// <summary>
     /// Effector is the interface for Casbin effectors.
@@ -10,10 +8,10 @@ namespace NetCasbin.Effect
         /// <summary>
         /// Merges all matching results collected by the enforcer into a single decision.
         /// </summary>
-        /// <param name="expr">The expression of [policy_effect].</param>
+        /// <param name="policyEffect">The expression of [policy_effect].</param>
         /// <param name="effects">The effects of all matched rules.</param>
         /// <param name="results">The matcher results of all matched rules.</param>
         /// <returns>The final effect.</returns>
-        bool MergeEffects(string expr, Effect[] effects, float[] results);
+        bool MergeEffects(string policyEffect, Effect[] effects, float[] results);
     }
 }