--- a/NetCasbin/Effect/PolicyEffectType.cs
+++ b/NetCasbin/Effect/PolicyEffectType.cs
@@ -1,0 +1,11 @@
+﻿namespace NetCasbin.Effect
+{
+    public enum PolicyEffectType
+    {
+        Custom,
+        AllowOverride,
+        AllowAndDeny,
+        DenyOverride,
+        Priority
+    }
+}