--- a/NetCasbin/InternalEnforcer.cs
+++ b/NetCasbin/InternalEnforcer.cs
@@ -387,7 +387,7 @@ namespace NetCasbin
             if (autoBuildRoleLinks && section.Equals(PermConstants.Section.RoleSection))
             {
                 BuildRoleLinks();
-                ExpressionProvider.SetGFunctions();
+                ExpressionHandler.SetGFunctions();
             }
 
             if (autoNotifyWatcher)
@@ -401,7 +401,7 @@ namespace NetCasbin
             if (autoBuildRoleLinks && section.Equals(PermConstants.Section.RoleSection))
             {
                 BuildRoleLinks();
-                ExpressionProvider.SetGFunctions();
+                ExpressionHandler.SetGFunctions();
             }
 
             if (autoNotifyWatcher && watcher is not null)